Cas de test enregistrer autour de la classe OperationMath

+++ src/main/java/com/jee/projetbleu/utils/OperationMath.java
@@ -3,7 +3,7 @@ package com.jee.projetbleu.utils;
 @javax.enterprise.context.RequestScoped
 public class OperationMath {
 
-    public float calculSimpleAddition(int value1, int value2) {
+    public float calculSimpleAddition(float value1, float value2) {
         return value1 + value2;
     }
